Refactored into separate test files for dymola, jmodelica and openmodelica
